Create process_log_input.py
This is the first script we use to process log data and dump into Redis. It treats the log entry itself as a "value" and we "score" the value with the computed epoch time taken from log entry timestamp. With ZSETs, it is easy to combine/sort log entries for a collection when more than one system is in play, since we are using the timestamp and not when we have "seen" the log entry.